Create codeql-analysis.yml
Commiting to enabling code scanning following guidance by @johnlugton @mario-campos @cmboling at GitHub Universe, as follows-
Go to the Code scanning alerts section in the Security tab.
Start the Set up this workflow step in the CodeQL Analysis card.
Review the created Action workflow and accept the default proposed workflow.
Head over to the Actions tab to see the created workflow in action.
You will notice the action failed because the Java job failed.
Reviewing a failed analysis job
CodeQL requires a build of compiled languages, and an analysis job can fail if our autobuilder is unable to build a program to extract an analysis database.

We are using this because we have a failed java job. Next steps- 
Head over to the failed Java job and try to determine the build failure.
Hint: The project is build using Maven so we can search for Failed to execute goal in the Autobuild step.
How would you resolve the build error to ensure the analysis can be performed?